content/media/webaudio/test/test_biquadFilterNode.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test BiquadFilterNode</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script src="webaudio.js" type="text/javascript"></script>
<script class="testbody" type="text/javascript">

function near(a, b, msg) {
  ok(Math.abs(a - b) < 1e-3, msg);
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  SpecialPowers.setBoolPref("media.webaudio.enabled", true);

  var context = new AudioContext();
  var buffer = context.createBuffer(1, 2048, 44100);
  for (var i = 0; i < 2048; ++i) {
    buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / 44100);
  }

  var destination = context.destination;

  var source = context.createBufferSource();

  var filter = context.createBiquadFilter();

  source.buffer = buffer;

  source.connect(filter);
  filter.connect(destination);

  // Verify default values
  is(filter.type, 0, "Correct default value for type");
  near(filter.frequency.minValue, 10, "Correct min value for filter frequency");
  near(filter.frequency.maxValue, 22050, "Correct max value for filter frequency");
  near(filter.frequency.defaultValue, 350, "Correct default value for filter frequency");
  near(filter.Q.minValue, 0.001, "Correct min value for filter Q");
  near(filter.Q.maxValue, 1000, "Correct max value for filter Q");
  near(filter.Q.defaultValue, 1, "Correct default value for filter Q");
  near(filter.gain.minValue, -40, "Correct min value for filter gain");
  near(filter.gain.maxValue, 40, "Correct max value for filter gain");
  near(filter.gain.defaultValue, 0, "Correct default value for filter gain");

  // Make sure that we can set all of the valid type values
  for (var i = 0; i <= 7; ++i) {
    filter.type = i;
  }

  expectException(function() {
    filter.type = 8;
  }, DOMException.INDEX_SIZE_ERR);

  source.start(0);
  SimpleTest.executeSoon(function() {
    source.stop(0);
    source.disconnect();
    filter.disconnect();

    SpecialPowers.clearUserPref("media.webaudio.enabled");
    SimpleTest.finish();
  });
});

</script>
</pre>
</body>
</html>
